Storing config and stickers in a different place now
