+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "doki-theme-hyper",
-  "version": "2.1.0",
+  "version": "2.1.1",
   "description": "The Doki Theme for Hyper",
   "main": "build/index.js",
   "repository": "https://github.com/Unthrottled/doki-theme-hyper.git",
