+++ src/settings.ts
@@ -29,7 +29,7 @@ const themes = Object.values(DokiThemeDefinitions)
     }
   });
 
-export const VERSION = 'v2.1.0';
+export const VERSION = 'v2.1.1';
 const icon = path.resolve(__dirname, '..', 'assets', 'Doki-Theme.png');
 const showAbout = () => {
   const appName = 'Doki Theme';
